--- src/main/java/com/spring/taskflow/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.controller;
-
-public class UserController {
-}
