+++ src/main/java/com/spring/taskflow/domain/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.controller;
+
+public class UserController {
+}
