+++ src/main/java/com/spring/taskflow/domain/dto/activitylog/ActivityLogListResponseDto.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.dto.activitylog;
+
+public class ActivityLogListResponseDto {
+}
