+++ src/main/java/com/spring/taskflow/domain/dto/comments/CommentsCreateRequestDto.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.dto.comments;
+
+public class CommentsCreateRequestDto {
+}
