+++ src/main/java/com/spring/taskflow/domain/dto/dashboard/DashboardCreateRequestDto.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.dto.dashboard;
+
+public class DashboardCreateRequestDto {
+}
