+++ src/main/java/com/spring/taskflow/domain/dto/tasks/TasksCreateRequestDto.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.dto.tasks;
+
+public class TasksCreateRequestDto {
+}
