+++ src/main/java/com/spring/taskflow/domain/dto/user/UserCreateRequestDto.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.dto.user;
+
+public class UserCreateRequestDto {
+}
