+++ src/main/java/com/spring/taskflow/domain/entity/Comments.java
@@ -1,4 +1,4 @@
-package com.spring.taskflow.entity;
+package com.spring.taskflow.domain.entity;
 
 import jakarta.persistence.*;
 import org.springframework.data.annotation.CreatedDate;
