+++ src/main/java/com/spring/taskflow/domain/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.repository;
+
+public interface UserRepository {
+}
