+++ src/main/java/com/spring/taskflow/domain/service/UserService.java
@@ -1,0 +1,4 @@
+package com.spring.taskflow.domain.service;
+
+public class UserService {
+}
