--- src/main/java/com/spring/taskflow/dto/activitylog/ActivityLogListResponseDto.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.dto.activitylog;
-
-public class ActivityLogListResponseDto {
-}
