--- src/main/java/com/spring/taskflow/dto/comments/CommentsCreateRequestDto.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.dto.comments;
-
-public class CommentsCreateRequestDto {
-}
