--- src/main/java/com/spring/taskflow/dto/dashboard/DashboardCreateRequestDto.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.dto.dashboard;
-
-public class DashboardCreateRequestDto {
-}
