--- src/main/java/com/spring/taskflow/dto/tasks/TasksCreateRequestDto.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.dto.tasks;
-
-public class TasksCreateRequestDto {
-}
