--- src/main/java/com/spring/taskflow/dto/user/UserCreateRequestDto.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.dto.user;
-
-public class UserCreateRequestDto {
-}
