--- src/main/java/com/spring/taskflow/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.repository;
-
-public interface UserRepository {
-}
