--- src/main/java/com/spring/taskflow/service/UserService.java
@@ -1,4 +1,0 @@
-package com.spring.taskflow.service;
-
-public class UserService {
-}
