+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
 spring.application.name=TaskFlow
 
 spring.datasource.url=jdbc:mysql://localhost:3306/taskflow
-spring.datasource.username=
-spring.datasource.password=
+spring.datasource.username=root
+spring.datasource.password=0000
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 spring.jpa.hibernate.ddl-auto=create
